// src/openapi3/rbac/1.0.0/models/rbac-v-100-target-rbac-role.ts
// GENERATED CODE -- DO NOT EDIT!
/* tslint:disable */
import { RbacV100TargetRbacRolePermission } from './rbac-v-100-target-rbac-role-permission';
export interface RbacV100TargetRbacRole {
  RbacV100targetRbacRolePermission?: RbacV100TargetRbacRolePermission;
  description?: string;
  roleid?: string;
}